+++ tests/cmds/test_profile.py
@@ -34,6 +34,18 @@ def mock_verify(mocker):
     return mocker.patch("code42cli.cmds.profile.validate_connection")
 
 
+@pytest.fixture
+def valid_connection(mock_verify):
+    mock_verify.return_value = True
+    return mock_verify
+
+
+@pytest.fixture
+def invalid_connection(mock_verify):
+    mock_verify.return_value = False
+    return mock_verify
+
+
 def test_show_profile_outputs_profile_info(capsys, mock_cliprofile_namespace, profile):
     profile.name = "testname"
     profile.authority_url = "example.com"
@@ -58,18 +70,6 @@ def test_show_profile_when_password_set_outputs_password_note(
     assert "A password is set" not in capture.out
 
 
-def test_create_profile_if_profile_exists_exits(capsys, mock_cliprofile_namespace):
-    mock_cliprofile_namespace.profile_exists.return_value = True
-    success = True
-    try:
-        profilecmd.create_profile("foo", "bar", "baz", True)
-    except SystemExit:
-        success = True
-        capture = capsys.readouterr()
-        assert "already exists" in capture.out
-    assert success
-
-
 def test_create_profile_if_user_sets_password_is_created(
     user_agreement, mock_verify, mock_cliprofile_namespace
 ):
@@ -86,7 +86,7 @@ def test_create_profile_if_user_does_not_set_password_is_created(
     mock_cliprofile_namespace.create_profile.assert_called_once_with("foo", "bar", "baz", True)
 
 
-def test_create_profile_if_user_does_not_set_password_does_not_save_password(
+def test_create_profile_if_user_does_not_agree_does_not_save_password(
     user_disagreement, mock_verify, mock_cliprofile_namespace
 ):
     mock_cliprofile_namespace.profile_exists.return_value = False
@@ -95,9 +95,8 @@ def test_create_profile_if_user_does_not_set_password_does_not_save_password(
 
 
 def test_create_profile_if_credentials_invalid_password_not_saved(
-    user_agreement, mock_verify, mock_cliprofile_namespace
+    user_agreement, invalid_connection, mock_cliprofile_namespace
 ):
-    mock_verify.return_value = False
     mock_cliprofile_namespace.profile_exists.return_value = False
     success = False
     try:
@@ -109,11 +108,57 @@ def test_create_profile_if_credentials_invalid_password_not_saved(
 
 
 def test_create_profile_if_credentials_valid_password_saved(
-    mocker, user_agreement, mock_verify, mock_cliprofile_namespace
+    mocker, user_agreement, valid_connection, mock_cliprofile_namespace
 ):
-    mock_verify.return_value = True
     mock_cliprofile_namespace.profile_exists.return_value = False
     profilecmd.create_profile("foo", "bar", "baz", True)
+    mock_cliprofile_namespace.set_password.assert_called_once_with("newpassword", mocker.ANY)
+
+
+def test_update_profile_updates_existing_profile(
+    mock_cliprofile_namespace, user_agreement, valid_connection, profile
+):
+    name = "foo"
+    profile.name = name
+    mock_cliprofile_namespace.get_profile.return_value = profile
+
+    profilecmd.update_profile(name, "bar", "baz", True)
+    mock_cliprofile_namespace.update_profile.assert_called_once_with(name, "bar", "baz", True)
+
+
+def test_update_profile_if_user_does_not_agree_does_not_save_password(
+    mock_cliprofile_namespace, user_disagreement, invalid_connection, profile
+):
+    name = "foo"
+    profile.name = name
+    mock_cliprofile_namespace.get_profile.return_value = profile
+    assert not mock_cliprofile_namespace.set_password.call_count
+
+
+def test_update_profile_if_credentials_invalid_password_not_saved(
+    user_agreement, invalid_connection, mock_cliprofile_namespace, profile
+):
+    name = "foo"
+    profile.name = name
+    mock_cliprofile_namespace.get_profile.return_value = profile
+
+    success = False
+    try:
+        profilecmd.create_profile("foo", "bar", "baz", True)
+    except SystemExit:
+        success = True
+        assert not mock_cliprofile_namespace.set_password.call_count
+    assert success
+
+
+def test_update_profile_if_user_agrees_and_valid_connection_sets_password(
+    mocker, user_agreement, valid_connection, mock_cliprofile_namespace, profile
+):
+    name = "foo"
+    profile.name = name
+    mock_cliprofile_namespace.get_profile.return_value = profile
+
+    profilecmd.update_profile(name, "bar", "baz", True)
     mock_cliprofile_namespace.set_password.assert_called_once_with("newpassword", mocker.ANY)