+++ CHANGELOG.md
@@ -10,7 +10,7 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 
 ## Unreleased
 
-### Changes
+### Changed
 
 - `securitydata` renamed to `security-data`.
 - From `security-data` related subcommands (such as `print`):
@@ -21,6 +21,7 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 
 ### Added
 
+- `code42 profile update` command.
 - `code42 profile create` command.
 - `code42 detection-lists high-risk` commands:
     - `bulk` with subcommands: